api-gateway/requirements.txt
requests
Flask==2.3.2
Flask-Limiter==2.7.0
PyJWT
Flask-CORS